Server:
-------

* Breaking change: No self-creation anymore
** App has to be included as package
** Must be created by factory method
** Config has to be passed to factory method
** Data source has to be passed to factory method
** Removed file for custom config

* Breaking change: Changed database interface
** Split database abstraction into classes for data source API and database adapters
*** Datasource classes: `AbstractDatasource`, `DefaultDatasource`
*** Database adapter classes: `AbstractDatabaseHandler`, `MysqlDatabaseHandler`
** Expecting new props `gameTitle` and `gameTag` in result of method `getClientIdentityData`
** Renamed database method `getSingleBotChannels` to `getGameBoundBotChannels`.
** Added new method `getGameStatuses` to read status for multiple games (used for bot commands)
** Removed methods `getGameStatus` and `getGameRound`
** Replaced config setting `SQL_DATABASE_PREFIX_GAME` in favor of `SQL_DATABASE_GAME`

* Breaking changes for usage of game bot:
** Changed command "galaxy?" to "game?"
** Removed command "round?", changed commands "game?" and "status?" to print all game info

* Breaking change: Webserver delivers own files by new request URLs
** Changed default webserver root from process working dir to package root dir
** Changed request URL for javascripts from "/js/..." to "/chat/webclient/..."
** Added request URL for webclient.js ("/chat/webclient.js")
** Removed partial files of webclient (formerly URL "/js/chat_controller.js", etc.)

* Mono-bot mode
** The mono-bot mode handles all game instances by exact one IRC bot
** (Instead of an IRC bot created per game instance)
** In-game chat users can be distinguished by added game info

* Various new config settings:
** `MAX_BOTS` can be used to configure mono-bot
** `WEB_SERVER_STATICS_DELIVERY_DIR` for own files
** `WEB_SERVER_CLIENT_DELIVERY_DIR` for files of the webclient (especially the `webclient.js`)
** `BOT_GAME_LABEL` for the common name of a game
** `SQL_DATABASE_GAME` in replace for `SQL_DATABASE_PREFIX_GAME`
** etc.

Webclient:
----------

* Breaking change: Reorganized webclient sources and requirements
** `webclient.js` now contains all sources, including thirdparty libraries
** Partial files of webclient are not provided any longer (`chat_controller.js`, etc.)
** Internal classes are hidden from the global space (except the main class)

* Breaking change: Renamed global main class `ChatController` to `GatewayChat`

* Breaking change: Target page must not contain the client's initial DOM / partial HTML anymore
** The webclient builds up the initial DOM by itself
** The DOM is appended to an existing HTML node, by default the body node
** Introduced new option `parentElement` to define the target HTML node

* Breaking change: Referencing and styling HTML nodes by CSS classes instead of IDs
** `#tabsystem` became `.tabsystem`
** `#tabPageServer` became `.tabPage.server`
** `#tabPage_<channel>` became `[data-id=tabPage_<channel>]`
** `#channelCreateForm` became `.channelCreateForm`
** `#channelNameInput` became `.channelNameInput`
** `#channelPasswordInput` became `.channelPasswordInput`
** `#channelFlagPublic` became `.channelFlagPublic`
** `#channelFlagIRC` became `.channelFlagIRC`
** `#channelCreateSubmitButton` became `.channelCreateSubmitButton`

* Breaking change: Further changes on CSS
** Add `.line` for messages having multiple lines
** Renamed `.tabsystemHeaders [data-id]` to `.tabsystemHeaders [data-tab]`

+++ src/server/botmanager.coffee
@@ -1,8 +1,16 @@
 
+## Include libraries
+Q = require 'q'
+
 ## Include app modules
-Config = require './config'
+config = require './config'
+log = require './logger'
+db = require './database'
+
+## Include app classes
 BotChannel = require './botchannel'
-Bot = require './bot'
+GameBot = require './gamebot'
+MonoBot = require './monobot'
 
 
 ## Abstraction of a service for watching existence of games.
@@ -11,23 +19,33 @@ Bot = require './bot'
 ##
 class BotManager
     isManaging: false
+    hasBotPerGame: false
     watcherTimer: null
 
     botList: null
     globalChannel: null
 
     constructor: ->
+        @hasBotPerGame = (config.MAX_BOTS > 0)
         @botList = {}
 
     start: =>
         # Setup bot channels
         globalChannelPromise = @_setupGlobalBotChannel()
-        singleChannelsPromise = @_setupSingleBotChannels()
+        globalChannelPromise.then (globalChannel) =>
+            @globalChannel = globalChannel
+        gameChannelsPromise = @_setupGameBoundBotChannels()
 
         # Setup the bots
-        botsPromise = @_setupBots()
+        botsPromise = @_destroyBots(@botList)  # Guard call: Destroy any bots created before
         botsPromise = botsPromise.then =>
-            return @_startBots(@botList)
+            if @hasBotPerGame
+                return @_setupBots()
+            else
+                return @_setupMonoBot()
+        botsPromise = botsPromise.then (botList) =>
+            @botList = botList  # Set new list of bots
+            return @_startBots(botList)
 
         # After bots startup: Add them to their channels
         botsPromise.then =>
@@ -36,17 +54,17 @@ class BotManager
             globalChannelPromise.then (globalChannel) =>
                 @_addBotsToGlobalChannel(globalChannel, @botList)
 
-            singleChannelsPromise.then (singleChannels) =>
-                @_addBotsToSingleChannels(singleChannels, @botList)
+            gameChannelsPromise.then (gameChannels) =>
+                @_addBotsToGameChannels(gameChannels, @botList)
 
             @_startGameWatcher()
 
         # End chain to observe errors
         botsPromise.done()
 
-        return Q.all([globalChannelPromise, singleChannelsPromise])
+        return Q.all([globalChannelPromise, gameChannelsPromise])
 
-    # Used by SocketHandler as callback to add a bot to a new channel 
+    # Used by SocketHandler as callback to add a bot to a new channel
     # (Therefor must be bound to BotManager instance)
     addGameBotToChannel: (gameID, channel) =>
         bot = @botList[gameID]
@@ -55,7 +73,12 @@ class BotManager
 
     shutdown: =>
         clearInterval(@watcherTimer) if @watcherTimer?
-        @_destroyBots(@botList)
+        promise = @_destroyBots(@botList)
+        promise = promise.then =>
+            return if @hasBotPerGame
+            bot = MonoBot.getInstance()
+            return bot.stopFinally()
+        return promise
 
 
     #
@@ -65,13 +88,35 @@ class BotManager
     _setupBots: ->
         promise = db.getBotRepresentedGames()
         promise = promise.then (gamesList) =>
-            for gameData in gamesList 
+            botList = {}
+
+            for gameData in gamesList
                 # Create bot
-                bot = new Bot(gameData)
+                bot = new GameBot(gameData)
 
                 # Store bot by game id
                 gameID = bot.getID()
-                @botList[gameID] = bot
+                botList[gameID] = bot
+
+            return botList
+
+        return promise
+
+    _setupMonoBot: ->
+        # Create the bot
+        bot = MonoBot.getInstance()
+
+        # Store mono-bot by each game id
+        promise = db.getBotRepresentedGames()
+        promise = promise.then (gamesList) =>
+            # Set initial list of games
+            bot.updateGamesList(gamesList)
+
+            # Store the bot for each game (makes handling easier)
+            botList = {}
+            for gameData in gamesList
+                botList[gameData.id] = bot
+            return botList
 
         return promise
 
@@ -79,12 +124,11 @@ class BotManager
         promise = db.getGlobalChannelData()
         promise = promise.then (channelData) =>
             log.info 'Creating global channel...'
-            @globalChannel = BotChannel.getInstance(channelData, true)
-            return @globalChannel
+            return BotChannel.getInstance(channelData, true)
         return promise
 
-    _setupSingleBotChannels: ->
-        promise = db.getSingleBotChannels()
+    _setupGameBoundBotChannels: ->
+        promise = db.getGameBoundBotChannels()
         promise = promise.then (channelList) =>
             log.info 'Creating additional bot channels...' if channelList?.length
             # Create every bot channel and push to result array
@@ -101,10 +145,11 @@ class BotManager
     #
 
     _startGameWatcher: ->
-        intervalFunc = => 
+        intervalFunc = =>
             @_manageBotsByGames()
-        timerMilliSeconds = Config.GAMES_LOOKUP_INTERVAL * 1000
-        @watcherTimer = setInterval(intervalFunc, timerMilliSeconds)
+        timerMilliSeconds = config.GAMES_LOOKUP_INTERVAL * 1000
+        clearInterval(@watcherTimer) if @watcherTimer?  # Clear old timer
+        @watcherTimer = setInterval(intervalFunc, timerMilliSeconds) if timerMilliSeconds > 0
 
     _manageBotsByGames: ->
         newBotsList = {}
@@ -112,7 +157,7 @@ class BotManager
 
         return if @isManaging
         @isManaging = true
-        log.debug 'Starting managing bots...'
+        log.debug 'Starting managing bots by games...'
 
         # Copy all old bots
         for key, bot of @botList
@@ -121,22 +166,33 @@ class BotManager
         # Check games
         promise = db.getBotRepresentedGames()
         promise = promise.then (gamesList) =>
-            for gameData in gamesList 
+            for gameData in gamesList
                 gameID = gameData.id
 
                 # Create bot of new game
                 unless @botList[gameID]?
-                    bot = new Bot(gameData)
+                    bot =
+                        if @hasBotPerGame
+                            new GameBot(gameData)
+                        else
+                            MonoBot.getInstance()
+
                     newBotsList[gameID] = bot
                     @botList[gameID] = bot
 
                 # Remove bot of existing game from end list
                 delete endBotsList[gameID]
+
+            # Only for mono-bot: Set list of games at once
+            unless @hasBotPerGame
+                bot = MonoBot.getInstance()
+                bot.updateGamesList(gamesList)
+
             return
 
         # Destroy bots to end
         promise = promise.then =>
-            return @_destroyBots(endBotsList)
+            return @_destroyBots(endBotsList, true)
 
         # Start new bots
         promise = promise.then =>
@@ -144,9 +200,9 @@ class BotManager
 
         # Join new bots to channels
         promise = promise.then =>
-            @_addBotsToGlobalChannel(@globalChannel, newBotsList)
+            @_addBotsToGlobalChannel(@globalChannel, newBotsList) if @globalChannel
             @isManaging = false
-            log.debug 'Finished managing bots!'
+            log.debug 'Finished managing bots by games!'
 
         # End chain to observe errors
         promise.done()
@@ -171,15 +227,15 @@ class BotManager
     _addBotsToGlobalChannel: (globalChannel, botList) ->
         # Add every bot to channel
         joinPromise = Q()
-        for key, bot of botList 
+        for key, bot of botList
             do (bot) =>
                 # Join bot as soon as previous bot has joined (To have clean logging order)
                 joinPromise = joinPromise.then =>
                     return @_addBotToChannel(bot, globalChannel)
 
-    _addBotsToSingleChannels: (singleChannels, botList) ->
+    _addBotsToGameChannels: (gameChannels, botList) ->
         # For each channel, add its appropriate single bot to it
-        for channel in singleChannels
+        for channel in gameChannels
             gameID = channel.getGameID()
             bot = botList[gameID]
             if bot?
@@ -193,34 +249,27 @@ class BotManager
         return channel.addBot(bot)
 
 
-    _destroyBots: (botList) ->
+    _destroyBots: (botList, deleteChannels=false) ->
         promises = for gameID, bot of botList
             do (gameID, bot) =>
-                destroy_promise = @_destroyBot(bot)
+                destroy_promise = @_destroyBot(bot, gameID)
                 destroy_promise.then =>
-                    db.deleteChannelsByGame(gameID)
+                    db.deleteChannelsByGame(gameID) if deleteChannels
                 return destroy_promise
         return Q.all(promises)
 
-    _destroyBot: (bot) ->
-        # Remove bot reference
-        delete @botList[bot.getID()]
+    _destroyBot: (bot, gameID) ->
+        # Remove bot reference in main list
+        delete @botList[gameID]
 
         # Remove bot from its channels (This is an optional soft shutdown action: Stopping the bot does the same on quit)
         promise = Q()
-        #promise = @_removeBotFromChannels(bot)
 
         # Finally disconnect the bot
         promise = promise.then =>
-            return bot.stop()
+            return bot.stop(gameID)
 
         return promise
-
-    _removeBotFromChannels: (bot) ->
-        channelList = bot.getWebChannelList()
-        promises = for key, channel of channelList
-            channel.removeBot(bot)
-        return Q.all(promises)
